Fix label losing "for" attribute and deprecation warnings.
When the input was re-rendered, the label's `for` attribute would
disappear (since the underlying elements all change). This is solved by
switching from `didInsertElement` (called only once as of Ember 1.13) to
`didRender` (called on re-renders too).

This also fixes a deprecation warning present due to the `for` attribute
being modified during this rendering phase by shifting it to the
`afterRender` callback.